Feat: rework readingView
readingView can now interact with the translate tab by clicking on
words. The pages functionality was also removed from the readingView,
as this will likely be handled on the db side of things later. Instead
of working with pages, readingView works directly with words. words is
an array of html words that regenerates based off of text upon switching
to readingView, this regeneration is handled by the genHtmWords
function.